requirements.txt
streamlit
google-ads
pandas